requirements.txt
PyGithub
boto
pymongo
thunder